[FIX] 네이밍 수정
