+++ src/components/Default/Wave/index.ts
@@ -1,0 +1,2 @@
+import Wave from './Wave';
+export default Wave;
